Rename: move class file
model 클래스와 test 코드 경로 수정
