+++ boogimonProject/src/main/java/model/UserJson.java
