--- boogimonProject/src/main/model/UserJson.json
@@ -1,1 +1,0 @@
-[{},{},{}]
